fix: repository 분리 오류 수정
