+++ src/main/java/com/example/yeobee/core/calculation/domain/CalculationRepository.java
@@ -14,4 +14,6 @@ public interface CalculationRepository {
     Long getTotalSharedBudgetExpense();
 
     List<CalculationResult> getTotalExpensePerTripUser(Long tripId, ExpenseType expenseType);
+
+    List<CalculationResult> getCalculationResult(Long tripId, ExpenseType expenseType);
 }
