+++ src/main/java/com/example/yeobee/core/expense/domain/CustomUserExpenseRepositoryImpl.java
@@ -2,19 +2,15 @@ package com.example.yeobee.core.expense.domain;
 
 import static com.example.yeobee.core.expense.domain.QExpense.expense;
 import static com.example.yeobee.core.expense.domain.QUserExpense.userExpense;
-import static com.example.yeobee.core.trip.domain.QTripCurrency.tripCurrency;
 import static com.example.yeobee.core.trip.domain.QTripUser.tripUser;
 
 import com.example.yeobee.common.exception.BusinessException;
 import com.example.yeobee.common.exception.ErrorCode;
-import com.example.yeobee.core.calculation.domain.CalculationResult;
 import com.example.yeobee.core.expense.dto.request.UserExpenseFilter;
 import com.example.yeobee.core.expense.dto.response.UserExpenseListRetrieveResponseDto;
 import com.querydsl.core.types.Predicate;
-import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.math.BigDecimal;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -49,25 +45,6 @@ public class CustomUserExpenseRepositoryImpl implements CustomUserExpenseReposit
                               pageable,
                               count);
     }
-
-    @Override
-    public List<CalculationResult> getCalculationResult(Long tripId, ExpenseType expenseType) {
-        return queryFactory.select(Projections.constructor(CalculationResult.class,
-                                                           tripUser,
-                                                           userExpense.amount
-                                                               .multiply(tripCurrency.exchangeRate.value)
-                                                               .divide(tripCurrency.exchangeRate.standard)
-                                                               .sum().coalesce(BigDecimal.ZERO)))
-            .from(userExpense)
-            .leftJoin(userExpense.expense, expense)
-            .leftJoin(userExpense.tripUser, tripUser)
-            .leftJoin(expense.tripCurrency, tripCurrency)
-            .where(expense.trip.id.eq(tripId)
-                       .and(expense.expenseType.eq(expenseType)))
-            .groupBy(tripUser.id)
-            .fetch();
-    }
-
 
     private Predicate[] getFilterPredicates(UserExpenseFilter filter) {
         return new Predicate[]{
